chore: break out common peripherals and rework peripheral init
Initialize common peripherals present on the CPU board in a
common peripherals library. Presently, this includes UART1, CAN and
CAN FD on SPI1.

This breaks the radio node functionality temporarily since UART1 is
configured for serial terminal. We will later move the radio node to an
IO board and use an analog port to receive SBUS data.