feat(sbus-receiver): use AN3 port to read SBUS data
Instead of using the CPU board's UART port, configure the AN3 port on
the IO board as a UART receive-only port. This change was made since the
CPU board's UART pins are connected directly to the CPU, while the IO
board's analog pins have a resistance between the pin and the CPU. This
protects the CPU from being burned, with the added bonus of being able
to print to console.