chore(battery-monitor): update cell voltage measurement for Rev C
The circuit used to measure individual cell voltages was updated due to
component unavailability. The new circuit measures the total voltage
over a range of cells instead of individual voltages. This requires us
to measure the cells in order and subtract each measured voltage by the
measured voltage of the previous cell.