feat: add subtrim support
Allows for setting a new neutral position for the servo/throttle by
using the trim buttons on the radio. The new neutral position will be
the position that's set when saving the calibration data. This way you
can calibrate the whole car from the radio. Calibration data persists
across reboot.