fix(fw_update): make all nodes enter the bootloader
Fixes bug where some nodes will not enter the bootloader after restart
when there is more than one node on the bus. Acheived by sending more
default letters. Also adds a sanity check to see if the nodes given in
the invocation of the script are present on the bus.