fix(spi-flash): add init workaround for cold reset issue
A timing issue was identified in the SPI flash memory on cold reset,
where the flash could not be read properly without first writing or
erasing a part of it. To address this, a new function
`spi_flash_workaround_init` has been introduced. This function attempts
to write to the last byte of the flash and, if unsuccessful, proceeds to
erase the last sector instead. Additionally, the LittleFS wrapper
configuration (`lfs-wrapper.c`) has been updated to reserve the last
sector for this initialization process, reducing the storage b 4KB.