fix(battery-monitor): prevent voltage spike on reg output turn on
When the reg output is off, the controller was trying to compensate for
a too low voltage by increasing the output voltage pot value. This makes
the pot have the maximum value when the controller is turned on,
resulting in a voltage spike. Fixed by returning early from functions
that depend on the controller being powered on to function.