fix(can-kingdom): handle ACTION_MODE_KEEP_CURRENT
Previously, the action mode was updated with the value of
ACTION_MODE_KEEP_CURRENT. Now, it is just ignored. The previous behavior
resulted in bugs in application code as the apps would start sending CAN
messages if they check for ACTION_MODE_FREEZE and
ACTION_MODE_KEEP_CURRENT is set.